api correction

+++ Minerva/ApiControllers/CommentController.cs
@@ -6,18 +6,19 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Net;
 using System.Net.Http;
+using System.Threading.Tasks;
 using System.Web.Http;
 
 namespace Minerva.ApiControllers
 {
+    //todo dodać [Authorize]
     public class CommentController : ApiController
     {
         private GenericRepository<MinervaDbContext, Comment, Int64> _commentRepository;
 
-        public CommentController(
-            GenericRepository<MinervaDbContext, Comment, Int64> commentRepository)
+        public CommentController()
         {
-            _commentRepository = commentRepository;
+            _commentRepository = new CommentRepository();
         }
 
         // GET: api/Comment
@@ -35,26 +36,54 @@ namespace Minerva.ApiControllers
         }
 
         // POST: api/Comment
-        public void Post([FromBody]Comment value)
+        public async Task<IHttpActionResult> Post([FromBody]Comment comment)
         {
-            throw new NotImplementedException();
+            if (!ModelState.IsValid)
+            {
+                return BadRequest(ModelState);
+            }
+
+            _commentRepository.Add(comment);
+            _commentRepository.Save();
+
+            return Ok();
         }
 
         // PUT: api/Comment/5
-        public void Put(int id, [FromBody]Comment value)
+        public async Task<IHttpActionResult> Put(int id, [FromBody]Comment comment)
         {
-            throw new NotImplementedException();
+            if (!ModelState.IsValid)
+            {
+                return BadRequest(ModelState);
+            }
+
+            var com = _commentRepository.FindBy(c => c.Id == comment.Id).FirstOrDefault();
+            
+            if (com == null)
+            {
+                return NotFound();
+            }
+
+            _commentRepository.Edit(comment);
+            _commentRepository.Save();
+
+            return Ok();
         }
 
         // DELETE: api/Comment/5
-        public void Delete(int id)
+        public async Task<IHttpActionResult> Delete(int id)
         {
             // todo ogranczenie co do własności
 
-            var entity = _commentRepository.FindBy(f => f.Id == id).First();
+            var entity = _commentRepository.FindBy(f => f.Id == id).FirstOrDefault();
+
+            if (entity == null)
+                return NotFound();
 
             _commentRepository.Delete(entity);
             _commentRepository.Save();
+
+            return StatusCode(HttpStatusCode.NoContent);
         }
     }
 }
